requirements.txt
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
docutils==0.20.1
google-re2==1.1
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
pkginfo==1.9.6
Pygments==2.16.1
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
six==1.16.0
sortedcontainers==2.4.0
twine==4.0.2
urllib3==2.0.4
webencodings==0.5.1
zipp==3.16.2
